refactor: Split Streamlit main orchestration and map prep modules
Refs: #243
